Fix the shortcuts for delete & goto bookmarks - now only one-key-pressing 'DELETE' or 'ENTER' works rather than any combinations. BUG-258
git-svn-id: http://svn.code.sf.net/p/genoviz/code/trunk@13775 80d5a9c2-7d8c-4c9f-9369-4e03bca4d2bd

+++ plugins/Bookmark/src/com/affymetrix/igb/bookmarks/BookmarkManagerView.java
@@ -243,7 +243,7 @@ public final class BookmarkManagerView implements TreeSelectionListener {
 	}
 
 	private void exportBookmarks(BookmarkList main_bookmark_list) {
-		if (main_bookmark_list == null) { // Support exporting from any node
+		if (main_bookmark_list == null) { // Support exporting from any node 
 			ErrorHandler.errorPanel("No bookmarks to save", (Exception) null, Level.SEVERE);
 			return;
 		}
@@ -510,6 +510,9 @@ public final class BookmarkManagerView implements TreeSelectionListener {
 
 			this.tree = tree;
 			this.tree.addKeyListener(new KeyListener() {
+				
+				// Set to remember which keys are pressed to handle combination keys pressed for deleting and adding bookmark
+				private final Set<Integer> pressedKeySet = new HashSet<Integer>(); 
 
 				@Override
 				public void keyTyped(KeyEvent e) {
@@ -517,16 +520,22 @@ public final class BookmarkManagerView implements TreeSelectionListener {
 
 				@Override
 				public void keyPressed(KeyEvent e) {
-					if ((e.getKeyCode() == KeyEvent.VK_DELETE)) {
-						deleteAction();
+					pressedKeySet.add(e.getKeyCode());
+					
+					if (pressedKeySet.size() ==1 && (e.getKeyCode() == KeyEvent.VK_DELETE)) { // Only if DELETE is pressed
+						deleteAction();	
+						pressedKeySet.remove(e.getKeyCode()); // Force remove because keyReleased() not always be triggered
 					}
-					if ((e.getKeyCode() == KeyEvent.VK_ENTER)) {
+					
+					if (pressedKeySet.size() ==1 && (e.getKeyCode() == KeyEvent.VK_ENTER)) { // Only if ENTER is pressed
 						goToAction();
+						pressedKeySet.remove(e.getKeyCode());
 					}
 				}
 
 				@Override
 				public void keyReleased(KeyEvent e) {
+					pressedKeySet.remove(e.getKeyCode());
 				}
 			});
 			this.def_tree_model = (DefaultTreeModel) tree.getModel();
